ユーザー情報更新時に画像情報が与えられたときのみS3の画像を削除するように修正

+++ app/controllers/api/v1/users_controller.rb
@@ -12,10 +12,12 @@ class Api::V1::UsersController < ApplicationController
   def update
     user = User.find(user_params[:id])
     client = Aws::S3::Client.new
-    client.delete_object({
-      bucket: ENV['S3_BUCKET'], 
-      key: user.image_key, 
-    })
+    if update_user_params.has_key?(:image_url) && update_user_params.has_key?(:image_key)
+      client.delete_object({
+        bucket: ENV['S3_BUCKET'], 
+        key: user.image_key, 
+      })
+    end
     user.update(update_user_params)
     render json: user, serializer: UserSerializer
   end
